import proptypes into components to reduce visual complexity. Use .npmignore to reduce npm package size.
