+++ .npmignore
@@ -1,0 +1,23 @@
+.DS_Store
+node_modules
+npm-debug.log
+coverage
+.nyc_output
+
+.bin
+demo
+mocha
+src
+.babelrc
+.editorconfig
+.eslintrc
+.nvmrc
+.travis.yml
+CodeBlock.js
+demo.js
+index.html
+README_DEVELOPMENT.md
+README_INTERACTION_TESTS.md
+webpack.config.demo.js
+webpack.config.react-responsive-select.js
+yarn.lock
