+++ src/components/SingleSelect.js
@@ -1,9 +1,8 @@
 import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import { SingleSelectProps } from '../propTypes';
 import singleline from 'singleline';
-import scrollIntoViewIIHOC from '../lib/scrollIntoViewIIHOC';
 import SingleSelectOption from './SingleSelectOption';
-
+import scrollIntoViewIIHOC from '../lib/scrollIntoViewIIHOC';
 const SingleSelectOptionHOC = scrollIntoViewIIHOC(SingleSelectOption);
 
 export default class SingleSelect extends Component {
@@ -112,35 +111,4 @@ export default class SingleSelect extends Component {
   }
 }
 
-SingleSelect.propTypes = {
-  altered: PropTypes.bool,
-  caretIcon: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.element
-  ]),
-  customLabelText: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.bool,
-    PropTypes.element
-  ]),
-  disabled: PropTypes.bool,
-  singleSelectInitialIndex: PropTypes.number,
-  singleSelectSelectedIndex: PropTypes.number,
-  singleSelectSelectedOption: PropTypes.shape({
-    text: PropTypes.string,
-    value: PropTypes.string
-  }),
-  isDragging: PropTypes.bool,
-  isOptionsPanelOpen: PropTypes.bool,
-  name: PropTypes.string,
-  nextPotentialSelectionIndex: PropTypes.number,
-  onSubmit: PropTypes.func,
-  options: PropTypes.arrayOf(
-    PropTypes.shape({
-      text: PropTypes.string.isRequired,
-      value: PropTypes.string.isRequired
-    })
-  ).isRequired,
-  prefix: PropTypes.string,
-  selectedValue: PropTypes.string
-};
+SingleSelect.propTypes = SingleSelectProps;
