+++ src/propTypes/index.js
@@ -1,0 +1,5 @@
+export { ReactResponsiveSelectProps } from './ReactResponsiveSelectProps';
+export { MultiSelectProps } from './MultiSelectProps';
+export { MultiSelectOptionProps } from './MultiSelectOptionProps';
+export { SingleSelectProps } from './SingleSelectProps';
+export { SingleSelectOptionProps } from './SingleSelectOptionProps';
